test(interaction): add faker spreadsheet helper
