+++ .github/workflows/emoji-helper.yml
@@ -1,4 +1,4 @@
-name: Emoji Helper
+name: 😄 Emoji Helper
 
 on:
   release:
