+++ .github/workflows/issue-close-require.yml
@@ -1,4 +1,3 @@
-
 name: Issue Close Require
 
 on:
