+++ packages/s2-core/__tests__/unit/interaction/base-interaction/click/data-cell-click-spec.ts
@@ -1,21 +1,16 @@
-import { Canvas, Event as GEvent } from '@antv/g-canvas';
-import EE from '@antv/event-emitter';
+import { createFakeSpreadSheet } from 'tests/util/helpers';
+import { Event as GEvent } from '@antv/g-canvas';
 import { omit } from 'lodash';
 import { DataCellClick } from '@/interaction/base-interaction/click';
 import { S2Options, ViewMeta } from '@/common/interface';
-import { Store } from '@/common/store';
 import { SpreadSheet } from '@/sheet-type';
-import { RootInteraction } from '@/interaction/root';
 import { InteractionStateName, S2Event } from '@/common/constant';
 
 jest.mock('@/interaction/event-controller');
 
-class FakeSpreadSheet extends EE {}
-
 describe('Interaction Data Cell Click Tests', () => {
   let dataCellClick: DataCellClick;
   let s2: SpreadSheet;
-  let interaction: RootInteraction;
   const mockCellViewMeta: Partial<ViewMeta> = {
     id: '1',
     colIndex: 0,
@@ -29,15 +24,12 @@ describe('Interaction Data Cell Click Tests', () => {
   };
 
   beforeEach(() => {
-    s2 = new FakeSpreadSheet() as unknown as SpreadSheet;
+    s2 = createFakeSpreadSheet();
     s2.getCell = () => mockCell as any;
-    s2.store = new Store();
-    interaction = new RootInteraction(s2 as unknown as SpreadSheet);
     dataCellClick = new DataCellClick(
       s2 as unknown as SpreadSheet,
-      interaction,
+      s2.interaction,
     );
-    s2.interaction = interaction;
     s2.options = {
       tooltip: {
         operation: {
@@ -45,11 +37,6 @@ describe('Interaction Data Cell Click Tests', () => {
         },
       },
     } as S2Options;
-    s2.container = {
-      draw: jest.fn(),
-    } as unknown as Canvas;
-    s2.hideTooltip = jest.fn();
-    s2.showTooltipWithInfo = jest.fn();
     s2.isTableMode = jest.fn(() => true);
   });
 
@@ -61,7 +48,7 @@ describe('Interaction Data Cell Click Tests', () => {
     s2.emit(S2Event.DATA_CELL_CLICK, {
       stopPropagation() {},
     } as unknown as GEvent);
-    expect(interaction.getState()).toEqual({
+    expect(s2.interaction.getState()).toEqual({
       cells: [mockCellMeta],
       stateName: InteractionStateName.SELECTED,
     });
