+++ packages/s2-core/__tests__/unit/interaction/base-interaction/click/row-column-click-spec.ts
@@ -1,6 +1,6 @@
 import { Canvas, Event as GEvent } from '@antv/g-canvas';
-import EE from '@antv/event-emitter';
 import { omit } from 'lodash';
+import { createFakeSpreadSheet } from 'tests/util/helpers';
 import { RowColumnClick } from '@/interaction/base-interaction/click';
 import {
   HiddenColumnsInfo,
@@ -8,7 +8,6 @@ import {
   S2Options,
   ViewMeta,
 } from '@/common/interface';
-import { Store } from '@/common/store';
 import { SpreadSheet } from '@/sheet-type';
 import { RootInteraction } from '@/interaction/root';
 import { InteractionStateName, S2Event } from '@/common/constant';
@@ -16,12 +15,9 @@ import { Node } from '@/facet/layout/node';
 
 jest.mock('@/interaction/event-controller');
 
-class FakeSpreadSheet extends EE {}
-
 describe('Interaction Data Cell Click Tests', () => {
   let rowColumnClick: RowColumnClick;
   let s2: SpreadSheet;
-  let interaction: RootInteraction;
   const mockCellViewMeta: Partial<ViewMeta> = {
     id: '1',
     colIndex: 0,
@@ -37,20 +33,16 @@ describe('Interaction Data Cell Click Tests', () => {
   };
 
   beforeEach(() => {
-    s2 = new FakeSpreadSheet() as unknown as SpreadSheet;
+    s2 = createFakeSpreadSheet();
     s2.getCell = () => mockCell as any;
-    s2.store = new Store();
-    s2.render = jest.fn();
-    interaction = new RootInteraction(s2 as unknown as SpreadSheet);
-    interaction.getActiveCells = () => [mockCell] as unknown as S2CellType[];
-    interaction.getRowColActiveCells = () =>
+    s2.interaction.getActiveCells = () => [mockCell] as unknown as S2CellType[];
+    s2.interaction.getRowColActiveCells = () =>
       [mockCell] as unknown as S2CellType[];
     rowColumnClick = new RowColumnClick(
       s2 as unknown as SpreadSheet,
-      interaction,
+      s2.interaction,
     );
     s2.isHierarchyTreeType = () => false;
-    s2.interaction = interaction;
     s2.options = {
       hiddenColumnFields: ['a'],
       tooltip: {
@@ -60,11 +52,6 @@ describe('Interaction Data Cell Click Tests', () => {
       },
     } as S2Options;
     s2.setOptions = jest.fn();
-    s2.container = {
-      draw: jest.fn(),
-    } as unknown as Canvas;
-    s2.hideTooltip = jest.fn();
-    s2.showTooltipWithInfo = jest.fn();
     s2.isTableMode = jest.fn(() => true);
   });
 
@@ -76,7 +63,7 @@ describe('Interaction Data Cell Click Tests', () => {
     s2.emit(S2Event.ROW_CELL_CLICK, {
       stopPropagation() {},
     } as unknown as GEvent);
-    expect(interaction.getState()).toEqual({
+    expect(s2.interaction.getState()).toEqual({
       cells: [mockCellMeta],
       nodes: [],
       stateName: InteractionStateName.SELECTED,
@@ -117,7 +104,7 @@ describe('Interaction Data Cell Click Tests', () => {
     // omit current expand node
     expect(s2.store.get('hiddenColumnsDetail')).toEqual([]);
     // reset interaction
-    expect(interaction.getState()).toEqual({
+    expect(s2.interaction.getState()).toEqual({
       cells: [],
       force: false,
     });
