+++ s2-site/docs/manual/advanced/interaction/custom.zh.md
@@ -64,7 +64,7 @@ s2.render();
 
 ## 3. 多个自定义交互
 
-当时，你可以注册任意数量的自定义交互，比如你不想让用户在表格右键时出现 浏览器默认的菜单
+当然，你可以注册任意数量的自定义交互，比如你不想让用户在表格右键时出现 浏览器默认的菜单
 
 ```ts
 import { TableSheet } from '@antv/s2';
